support base64 path

+++ files.js
@@ -2,8 +2,10 @@ const fs = require("fs");
 const http = require("http");
 const https = require("https");
 const fsExtra = require("fs-extra");
+const path = require("path");
 
 const downloadImagesAndSaveToFile = async (directory, images) => {
+  console.log("download images...");
   const imgDir = `${directory}/images`;
 
   // Delete old files
@@ -14,29 +16,43 @@ const downloadImagesAndSaveToFile = async (directory, images) => {
   }
 
   for (let i = 0; i < images.length; i++) {
-    result = await download(images[i], `${imgDir}/image-${i}.png`);
+    const ext = path.extname(images[i].src);
+    result = await download(images[i].src, `${imgDir}/image-${i}${ext}`);
   }
+
+  console.log("download images end.");
 };
 
 const download = (url, destination) => {
-  const proto = !url.charAt(4).localeCompare("s") ? https : http;
-  return new Promise((resolve, reject) => {
-    const file = fs.createWriteStream(destination);
+  if (url.indexOf("data:") !== -1) {
+    const ReadableData = require("stream").Readable;
+    const imageBufferData = Buffer.from(url, "base64");
+    const streamObj = new ReadableData();
+    streamObj.push(imageBufferData);
+    streamObj.push(null);
+    streamObj.pipe(fs.createWriteStream(destination));
+    return;
+  } else {
+    const proto = !url.charAt(4).localeCompare("s") ? https : http;
 
-    proto
-      .get(url, (response) => {
-        response.pipe(file);
+    return new Promise((resolve, reject) => {
+      const file = fs.createWriteStream(destination);
 
-        file.on("finish", () => {
-          file.close(resolve(true));
+      proto
+        .get(url, (response) => {
+          response.pipe(file);
+
+          file.on("finish", () => {
+            file.close(resolve(true));
+          });
+        })
+        .on("error", (error) => {
+          fs.unlink(destination);
+
+          reject(error.message);
         });
-      })
-      .on("error", (error) => {
-        fs.unlink(destination);
-
-        reject(error.message);
-      });
-  });
+    });
+  }
 };
 
 const initDirectory = (directory) => {
@@ -51,7 +67,7 @@ const copyFilesToDirectory = async (directory) => {
     `${directory}/style.css`,
     (err) => {
       if (err) {
-        console.log("error", err);
+        console.log("Can't copy styles", err);
       }
     }
   );
